Added test files

+++ src/data/Bool.java
@@ -66,7 +66,7 @@ public class Bool extends Data {
     @Override
     public Data quotientOperator(Data secondTerm)  throws AmlRunTimeException {
         if(secondTerm instanceof Int) {
-            if (secondTerm.getValue() == 0) throw new AmlRunTimeException("Floating point exception: Division by 0.");
+            if ((int)secondTerm.getValue() == 0) throw new AmlRunTimeException("Floating point exception: Division by 0.");
             return new Int(toInt() / ((Int) secondTerm).getValue());
         } else if(secondTerm instanceof Bool) {
             int value = ((Bool) secondTerm).toInt();
