+++ src/data/Int.java
@@ -61,7 +61,7 @@ public class Int extends Data {
     @Override
     public Data quotientOperator(Data secondTerm) throws AmlRunTimeException {
         if(secondTerm instanceof Int) {
-            if (secondTerm.getValue() == 0) throw new AmlRunTimeException("Floating point exception: Division by 0.");
+            if ((int)secondTerm.getValue() == 0) throw new AmlRunTimeException("Floating point exception: Division by 0.");
             return new Int(dataInt / ((Int) secondTerm).getValue());
         } else if(secondTerm instanceof Bool) {
             int value = ((Bool) secondTerm).toInt();
